+++ src/explorer/model.ts
