+++ src/webview/donate.ts
@@ -1,5 +1,5 @@
 import { ExtensionContext, ViewColumn } from 'vscode';
-import ReusedWebviewPanel from '../ReusedWebviewPanel';
+import ReusedWebviewPanel from './ReusedWebviewPanel';
 
 async function donate(context: ExtensionContext) {
   const panel = ReusedWebviewPanel.create('donateWebview', '打赏作者@giscafer', ViewColumn.One, {
