+++ src/webview/stockTrend.ts
@@ -1,5 +1,5 @@
 import { ViewColumn } from 'vscode';
-import ReusedWebviewPanel from '../ReusedWebviewPanel';
+import ReusedWebviewPanel from './ReusedWebviewPanel';
 import stockTrendPic from './stockTrendPic';
 
 function stockTrend(code: string, name: string, stockCode: string) {
